Improve game logic in a different direction
Started to implement just only one click handler to handle to game process. Simplified the logic calls to be in only one event listener, changing control paths based on the image clicked.
Currently the functionality supports a click limit, and this implementation solves the problem of clicking too fast and getting an incorrect score at the end of the game sequence.
This is done by removing the event listener when the score limit has been reached.
More implementation will be needed to regain functionality of the game in this branch version.